bpa-online/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-dateutil>=2.8.2
pandas>=2.2.0
psycopg2-binary>=2.9.9
unidecode>=1.3.7
python-dotenv>=1.0.0
PyJWT>=2.8.0
requests>=2.31.0
urllib3>=2.1.0
dbfread>=2.0.7
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
bcrypt>=4.1,<5.0
gunicorn>=21.2,<22.0


